backend/functions/requirements.txt
firebase_functions~=0.1.0
openai==0.28.1
python-dotenv==1.0.0